// MercaditoTec/src/app/models/course-tutor-interface.ts
export interface CourseTutorInterface {
    idCursoTutor: number;
    idTutor: number;
    idEstudiante: number;
    nombre: string;
    apellidos: string;
    calificacionPromedioTutor: number;
    idCurso: number;
    nombreCurso: string;
    codigo: string;
    carrera: string;
    notaObtenida: number;
    temas: string;
}